update to samples v2
