+++ sdk/web-pubsub/web-pubsub/sample.env
@@ -1,0 +1,1 @@
+WPS_CONNECTION_STRING="<connection string>"
