+++ sdk/web-pubsub/web-pubsub/samples/v1/javascript/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "azure-web-pubsub-samples-js",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Azure Web PubSub client library samples for JavaScript",
+  "engine": {
+    "node": ">=12.0.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
+    "directory": "sdk/web-pubsub/web-pubsub"
+  },
+  "keywords": [
+    "Azure",
+    "cloud"
+  ],
+  "author": "Microsoft Corporation",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
+  },
+  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/web-pubsub/web-pubsub",
+  "dependencies": {
+    "@azure/web-pubsub": "next",
+    "dotenv": "latest"
+  }
+}
