+++ sdk/web-pubsub/web-pubsub/samples/v1/javascript/sample.env
@@ -1,0 +1,1 @@
+WPS_CONNECTION_STRING="<connection string>"
