+++ sdk/web-pubsub/web-pubsub/samples/v1/typescript/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "azure-web-pubsub-samples-ts",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Azure Web PubSub client library samples for TypeScript",
+  "engine": {
+    "node": ">=12.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prebuild": "rimraf dist/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
+    "directory": "sdk/web-pubsub/web-pubsub"
+  },
+  "keywords": [
+    "Azure",
+    "cloud"
+  ],
+  "author": "Microsoft Corporation",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
+  },
+  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/web-pubsub/web-pubsub",
+  "dependencies": {
+    "@azure/web-pubsub": "next",
+    "dotenv": "latest"
+  },
+  "devDependencies": {
+    "typescript": "~4.2.0",
+    "rimraf": "latest"
+  }
+}
