+++ sdk/web-pubsub/web-pubsub/samples/v1/typescript/src/broadcasting.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * @summary Demonstrates broadcasting messages to a hub and a group
+ */
+
+import { WebPubSubServiceClient } from "@azure/web-pubsub";
+
+import * as dotenv from "dotenv";
+dotenv.config();
+
+const chatHub = new WebPubSubServiceClient(process.env.WPS_CONNECTION_STRING!, "chat");
+
+async function main() {
+  // send a text message to the entire hub
+  await chatHub.sendToAll("Hi there!");
+
+  // send a text message to a particular group
+  const adminGroup = chatHub.group("admin");
+  await adminGroup.sendToAll("Hi admins!");
+
+  // send binary data to the entire hub
+  const data = new Uint8Array(10);
+  for (let i = 0; i < 10; i++) {
+    data[i] = i;
+  }
+  chatHub.sendToAll(data.buffer);
+}
+
+main();
